Implemented add movie endpoint for the movie controller

+++ src/main/java/com/tolgaziftci/casestudy/controllers/MovieController.java
@@ -1,7 +1,10 @@
 package com.tolgaziftci.casestudy.controllers;
 
+import com.tolgaziftci.casestudy.dto.AddMovieDTO;
+import com.tolgaziftci.casestudy.exceptions.MovieAlreadyExistsException;
 import com.tolgaziftci.casestudy.exceptions.MovieNotFoundException;
 import com.tolgaziftci.casestudy.models.Movie;
+import com.tolgaziftci.casestudy.services.MappingService;
 import com.tolgaziftci.casestudy.services.MovieService;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,9 +14,11 @@ import java.util.List;
 @RequestMapping("/api")
 public class MovieController {
     private final MovieService movieService;
+    private final MappingService mappingService;
 
-    public MovieController(MovieService movieService) {
+    public MovieController(MovieService movieService, MappingService mappingService) {
         this.movieService = movieService;
+        this.mappingService = mappingService;
     }
 
     @GetMapping("/movies")
@@ -28,8 +33,19 @@ public class MovieController {
         else throw new MovieNotFoundException(id);
     }
 
+    @PostMapping("/movies")
+    public Movie addMovie(@RequestBody AddMovieDTO dto){
+        if (movieService.getMovieByTitle(dto.getTitle()) == null){
+            Movie movie = mappingService.convertDTOToMovie(dto);
+            movieService.addMovie(movie);
+            return movie;
+        }
+        else throw new MovieAlreadyExistsException(dto.getTitle());
+    }
+
     @DeleteMapping("/movies/{id}")
     public void deleteMovie(@PathVariable("id") Integer id) {
-        movieService.deleteMovie(movieService.getMovieById(id));
+        if (movieService.getMovieById(id) != null) movieService.deleteMovie(movieService.getMovieById(id));
+        else throw new MovieNotFoundException(id);
     }
 }
